Backend/deployment-skin/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-storage
azure-storage-blob
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.3.0.36
opencv-python==4.2.0.32
Pillow==7.2.0
Pyrebase4==4.4.3
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
tensorflow-cpu==2.3.1
uritemplate==3.0.1
urllib3==1.25.9
validators==0.18.2
wasabi==0.7.1
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0